Add support for creating new projects with existing files
We also recognize and handle the following existing entries in the project
root:

  .git      -- assume VCS is already initialized and is Git

  LICENSE   -- try to guess the license from contents

  README.md -- try to extract the summary line from contents

Overall, the idea is to streamline the workflow where one creates a project on
one of the hosting services (GitHub, GitLab) and then initializes it with
bdep-new.

Also, to this effect, specifying the project name is now optional and if
omitted the current working directory (or --output-dir, if specified) is
assumed to be the project name.

Here is an example of this workflow:

$ # create project with LICENSE and README.md on Git hosting service
$ git clone .../libhello.git
$ cd libhello
$ bdep new -t lib