Add --no-sync option to bdep-init
This allows postponing initialization in the build configurations to a later
explicit bdep-sync. One subtle difference with such an explicit sync is that
it will be performed without having the project database open, which can be
important if our initialization triggers an implicit sync (via a hook) of a
project that uses the same database (as is the case with build system module
projects).