Add third-party bdep-new type sub-option for libraries and executables
This sub-option is meant for converting an existing third-party project
to build2. It automatically enables a number of other sub-options (such
as no-version, no-readme, and no-symexport). It also adds a number of
values to manifest that makes sense to specify in a package of a third-
party project and, unless no-package-readme is specified, generates the
PACKAGE-README.md template.